+++ src/ui/components/obsidian/tree.svelte
@@ -9,11 +9,15 @@
     title,
     flair,
     controls,
+    onpointermove,
+    onpointerup,
   }: {
     title: string;
     children: Snippet;
     flair?: Snippet;
     controls?: Snippet;
+    onpointermove?: (event: PointerEvent) => void;
+    onpointerup?: (event: PointerEvent) => void;
   } = $props();
 
   let isTreeVisible = $state(true);
@@ -28,7 +32,7 @@
 </script>
 
 <!--Partially uses Obsidian's classes for search result matches-->
-<div class="tree-container">
+<div class="tree-container" {onpointermove} {onpointerup}>
   <div class="tree-header-container">
     <div class="tree-item-self is-clickable" onclick={toggleTree}>
       <div