+++ src/ui/components/timeline-with-controls.svelte
@@ -52,31 +52,29 @@
   <TimelineControls />
 
   {#if $settings.showUncheduledTasks}
-    <Tree title="All day events">
+    <Tree
+      onpointermove={handleResizeableBoxPointerMove}
+      onpointerup={editContext.confirmEdit}
+      title="All day events"
+    >
       {#snippet flair()}
-        {String(displayedAllDayTasks.length)}
+        {#if editOperation.current}
+          Drag here to schedule all-day events
+        {:else}
+          {String(displayedAllDayTasks.length)}
+        {/if}
       {/snippet}
-      {#if editOperation.current || displayedAllDayTasks.length > 0}
-        <ResizeableBox
-          class="unscheduled-task-container"
-          onpointermove={handleResizeableBoxPointerMove}
-          onpointerup={editContext.confirmEdit}
-        >
+      {#if displayedAllDayTasks.length > 0}
+        <ResizeableBox class="unscheduled-task-container">
           {#snippet children(startEdit)}
-            {#if editOperation.current && displayedAllDayTasks.length === 0}
-              <div class="edit-prompt">
-                Drag blocks here to schedule all-day tasks
-              </div>
-            {:else if displayedAllDayTasks.length > 0}
-              <BlockList list={displayedAllDayTasks}>
-                {#snippet match(task: Task)}
-                  <UnscheduledTimeBlock
-                    --time-block-padding="var(--size-2-1) 0"
-                    {task}
-                  />
-                {/snippet}
-              </BlockList>
-            {/if}
+            <BlockList list={displayedAllDayTasks}>
+              {#snippet match(task: Task)}
+                <UnscheduledTimeBlock
+                  --time-block-padding="var(--size-2-1) 0"
+                  {task}
+                />
+              {/snippet}
+            </BlockList>
             <ResizeHandle on:mousedown={startEdit} />
           {/snippet}
         </ResizeableBox>
@@ -131,17 +129,6 @@
 
   :global(.unscheduled-task-container) {
     overflow: auto;
-  }
-
-  .edit-prompt {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    padding: var(--size-4-2);
-
-    font-size: var(--font-ui-small);
-    color: var(--text-faint);
   }
 
   .control-text {