+++ src/util/task-utils.ts
@@ -621,4 +621,8 @@ function getIndentedText(
     return result + "\n" + getIndentedText(current, indentation, options);
   }, listItemLineWithParagraphs);
 }
+
+export function isCompleted(taskCheckmark?: string) {
+  return taskCheckmark !== undefined && taskCheckmark.toLowerCase() === "x";
+}
 /* eslint-enable @typescript-eslint/no-floating-promises, @typescript-eslint/no-misused-promises, @typescript-eslint/no-redundant-type-constituents, @typescript-eslint/no-unnecessary-condition, @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-enum-comparison, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-return -- Re-enable scorecard compatibility suppressions after this file. */